rest

+++ src/main/java/kz/edu/service/interfaces/IEntityService.java
@@ -6,6 +6,6 @@ public interface IEntityService<T> {
     void add(T entity);
     void update(T entity);
     List<T> getAll();
-    T getById(int id);
-    void delete(int id);
+    T getById(long id);
+    void delete(long id);
 }
